Reorganiser les fixtures pytest
Mise en commun des fixtures qui sont utilisables en dehors de l'app SV.
